Animation: Easing Test
I know, it is not a test in the classic sense.

But testing easing is a matter of visually inspecting the curves.

I am sure that Visualization experts will concur...

It creates python scripts that visualize the easing functions

with matplotlib. All actual computations are done in Inviwo::Easing.



Found some typos using this

and could verify the correctness of all easing functions.