BaseGL: Fix to ensure that regular event is propagated if picking event is not used
CanvasOpengGLWidget::propagateEvent will first propagate the picking event and then the regular event (if the picking event was not used). The regular event will not be propagated past the viewmanager if it is already is marked as visited when passed along with the picking event. This causes problems for processors not handling the picking, but does handle the original event. E.g., when using EventMatcher since it does not match the picking event, but matches other events. Therefore, the original event must also be propagated to processors up the chain.